sistemazioni varie, manca db

+++ src/core/model.ts
@@ -2,20 +2,26 @@ import { Persistence } from "src/repository/persistence"
 import { Dynamo } from "src/repository/dynamo"
 import Address from "./address";
 import { DbMock } from "src/repository/dbMock";
+import { Users } from "src/repository/users";
+import { UsersService } from "src/repository/usersService";
+import { UsersMock } from "src/repository/usersMock";
 
 export class Model {
     private readonly DATABASE: Persistence;
+    private readonly USERS: Users;
     
-    private constructor (db: Persistence) {
+    private constructor (db: Persistence,  users: Users) {
         this.DATABASE = db;
+        this.USERS = users;
     }
 
     public static createModel (): Model {
-        return new Model(new Dynamo());
+        return new Model(new Dynamo(), new UsersService());
+        
     }
 
     public static createModelMock (): Model {
-        return new Model(new DbMock());
+        return new Model(new DbMock(), new UsersMock());
     } 
 
     /**
@@ -23,7 +29,11 @@ export class Model {
      * @param data the data of the address to add 
      * @returns the result of the operation
      */
-    public createAddress (data: {[key:string]: any} ): boolean {
+    public async createAddress (data: {[key:string]: any}, token: string ): Promise<boolean> {
+        const IS_VENDOR = await this.USERS.checkVendor(token);
+        if (!IS_VENDOR){
+            throw new Error("invalid token");
+        }
         let result = false;
         if(data) {
             const ADDRESS = new Address(data);
@@ -37,16 +47,12 @@ export class Model {
      * @param user the id of the user
      * @returns the addresses of the user with the given id
      */
-    public getAddresses (user: string): JSON {
-        const ADDRESSES: Array<Address> = this.DATABASE.getAll(user);
+    public async getAddresses (token: string): Promise<JSON> {
+        const USER = await this.USERS.checkUser(token);
+        const ADDRESSES: Array<Address> = this.DATABASE.getAll(USER);
         if(ADDRESSES == null)
             return null;
-        
-        const OBJ = [];
-        ADDRESSES.forEach(item => {
-            OBJ.push(item)
-        });
-        return JSON.parse(JSON.stringify(OBJ));
+        return JSON.parse(JSON.stringify(ADDRESSES));
     }
 
     /**
@@ -64,7 +70,14 @@ export class Model {
      * @param id the id of the address to delete
      * @returns the result of the operation
      */
-    public deleteAddress (id: string): boolean {
+    public async deleteAddress (id: string, token: string): Promise<boolean> { 
+        const IS_VENDOR = await this.USERS.checkVendor(token);
+        if (!IS_VENDOR){
+            throw new Error("invalid token");
+        }
+        if(!id){
+            return false;
+        }
         return this.DATABASE.deleteItem(id);
     }
 
@@ -74,7 +87,11 @@ export class Model {
      * @param data the data to modify
      * @returns the result of the operation
      */
-    public updateAddress (data: JSON): boolean {
+    public async updateAddress (token: string, data: JSON): Promise<boolean> {
+        const IS_VENDOR = await this.USERS.checkVendor(token);
+        if (!IS_VENDOR){
+            throw new Error("invalid token");
+        }
         if(!data)
             return false;
         return this.DATABASE.editItem(data);
